Cast to user-configured dtype in XarrayDataset (#1159)
Currently we cast to `torch.float` in`SingleModuleStepper.run_on_batch`.
This is a bit of an unexpected side-effect of `run_on_batch`. Also,
because it does not happen in `SingleModuleStepper.predict`, there can
be surprising differences between training/evaluator (which use
run_on_batch) and inference (which uses predict). E.g. I hit this when
running inference with an ERA5 model that has double precision CO2 in
the dataset, which ended up failing even though training worked on the
same dataset.

This PR makes three changes:

1) cast data in `XarrayDataset.__getitem__` instead of in
`run_on_batch`. This makes it consistent with the horizontal and
vertical coordinates, which were already being cast by `XarrayDataset`.
2) allow user to specify what dtype to cast to (or use None to not cast
at all). Default is `float32`, maintaining current behavior.
3) rename `cast_tensordict_to_device` to `move_tensordict_to_device`
since now this only needs to have one responsibility.

+++ fme/fme/core/data_loading/data_typing.py
@@ -147,8 +147,8 @@ class LatLonCoordinates(HorizontalCoordinates):
     def coords(self) -> Mapping[str, np.ndarray]:
         # TODO: Replace with lat/lon name?
         return {
-            "lat": self.lat.cpu().numpy(),
-            "lon": self.lon.cpu().numpy(),
+            "lat": self.lat.cpu().type(torch.float32).numpy(),
+            "lon": self.lon.cpu().type(torch.float32).numpy(),
         }
 
     @property
@@ -214,9 +214,9 @@ class HEALPixCoordinates(HorizontalCoordinates):
     @property
     def coords(self) -> Mapping[str, np.ndarray]:
         return {
-            "face": self.face.cpu().numpy(),
-            "height": self.height.cpu().numpy(),
-            "width": self.width.cpu().numpy(),
+            "face": self.face.cpu().type(torch.float32).numpy(),
+            "height": self.height.cpu().type(torch.float32).numpy(),
+            "width": self.width.cpu().type(torch.float32).numpy(),
         }
 
     @property
